+++ tomcat/src/main/java/org/apache/coyote/http11/http11response/ResponseManager.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Objects;
 import org.apache.coyote.http11.ExtensionContentType;
-import org.apache.coyote.http11.StatusCode;
 
 public class ResponseManager {
